rom: read NES 2.0 PRG and CHR size MSBs from byte 9
A NES 2.0 header keeps PRG size bits 8-11 in the low nibble of byte 9
and CHR size bits 8-11 in the high nibble; a nibble of $F instead makes
the size byte an exponent and multiplier, 2^E * (2M+1) bytes. The
loader ignored byte 9, so a 4 MB PRG ROM loaded as 0 bytes and odd
sizes such as 8 KB PRG could not be expressed.

Both forms are now decoded, and sizes above 64 MB are refused rather
than wrapped. Mapper.prg_banks and chr_banks grow to 16 bits so more
than 255 banks count correctly. That changes sizeof(NES) again, which
the header's struct size check covers; the state version was already
raised to 3 for the FME-7/MMC5 register move.

+++ src/nes/mapper.h
@@ -66,7 +66,7 @@ typedef struct Mapper {
     /* PRG ROM banking */
     uint8_t *prg_rom;
     uint32_t prg_rom_size;
-    uint8_t prg_banks;      /* Number of 16KB PRG banks */
+    uint16_t prg_banks;     /* Number of 16KB PRG banks */
     uint8_t prg_bank0;      /* Bank at $8000-$BFFF */
     uint8_t prg_bank1;      /* Bank at $C000-$FFFF */
     uint8_t prg_mode;       /* PRG banking mode */
@@ -74,7 +74,7 @@ typedef struct Mapper {
     /* CHR ROM/RAM banking */
     uint8_t *chr_rom;
     uint32_t chr_rom_size;
-    uint8_t chr_banks;      /* Number of 8KB CHR banks (or 4KB for some mappers) */
+    uint16_t chr_banks;     /* Number of 8KB CHR banks (or 4KB for some mappers) */
     uint8_t chr_bank0;      /* Bank at $0000-$0FFF (or $0000-$1FFF) */
     uint8_t chr_bank1;      /* Bank at $1000-$1FFF */
     uint8_t chr_ram[0x2000]; /* CHR RAM for mappers that use it */
